renamed files to better follow convention
